refactor: Extract TaskClient and ProjectClient sub-clients
Split the task and project endpoints out of the monolithic `TodoistApi`
class (3455 lines, 113 methods) into dedicated internal sub-clients as a
pilot for broader domain decomposition.

`TodoistApi` now instantiates `TaskClient` and `ProjectClient` as private
fields in its constructor; each public task/project method is a one-line
delegation. Public API, JSDoc, and test files are unchanged — all 479
tests pass without modification.

- New `src/clients/base-client.ts` holds the shared HTTP identity
  (authToken, syncApiBase, customFetch) and a `syncContext` accessor
- `src/transport/sync-request.ts` extracts the private `requestSync`
  method into a plain `performSyncRequest` utility so every sub-client
  (and the remaining methods on TodoistApi) can use it
- `src/utils/request-helpers.ts` lifts shared `generatePath` and
  `spreadIfDefined` helpers previously local to todoist-api.ts

Follow-up PRs will migrate the remaining ~15 domains (sections, labels,
comments, reminders, workspaces, folders, uploads, etc.) using the same
pattern.
